% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PH.Transfer.CV.fit}
\alias{PH.Transfer.CV.fit}
\title{Cox proportional hazards model with adaptive information transfer}
\usage{
PH.Transfer.CV.fit(
  yobs,
  delta,
  X,
  aux,
  uncertain = list(do = TRUE, haveV = TRUE),
  hetero = list(do = TRUE, group = TRUE),
  Lams = list(tm = NULL),
  eps = 1e-06,
  maxit = 50000,
  trace = TRUE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{aux}{indicates the aggregate statistics extracted from external sources (based on transitional models). It is a list of lists, and each sub-list represents information from a study.}

\item{uncertain}{a list that contains two elements:
\code{do} is a logical to indicate whether the uncertainty issue should be addressed;
\code{haveV} is another logical value to show whether we have external estimated variance-covariance matrices.}

\item{hetero}{a list that contains two elements:
\code{do} is a logical to indicate whether the heterogeneity issue should be addressed;
\code{group} is another logical value to show whether the group Lasso penalty should be used.}

\item{Lams}{a list that contains \code{tm}, which denotes a vector time points where it will evaluate the baseline cumulative hazard function on.}

\item{eps}{the error of convergence.}

\item{maxit}{the maximum number of iteration.}

\item{trace}{a logical to indicate whether the traces of model fitting should be printed.}
}
\description{
Fit the Cox proportional hazards model with aggregate estimate from multiple external sources.
}
\examples{
# Here is an example (to appear).

}
